[meta] fix building with clang 21.1.4 (#2839)
Reviewed-on: https://git.eden-emu.dev/eden-emu/eden/pulls/2839
Co-committed-by: Caio Oliveira <caiooliveirafarias0@gmail.com>

+++ src/core/hle/kernel/k_thread.h
@@ -1,3 +1,6 @@
+// SPDX-FileCopyrightText: Copyright 2025 Eden Emulator Project
+// SPDX-License-Identifier: GPL-3.0-or-later
+
 // SPDX-FileCopyrightText: Copyright 2021 yuzu Emulator Project
 // SPDX-License-Identifier: GPL-2.0-or-later
 
@@ -539,7 +542,7 @@ public:
     }
 
     void ClearDpc(DpcFlag flag) {
-        this->GetStackParameters().dpc_flags &= ~static_cast<u8>(flag);
+        this->GetStackParameters().dpc_flags &= static_cast<u8>(~static_cast<u8>(flag));
     }
 
     u8 GetDpc() const {
